fix(settings): per-device badge, PWA-only notifications, divider, inline % w/ decimals
1a) App-icon badge opt-in is now PER-DEVICE (localStorage), not account-wide —
    two phones can independently show/hide a budget's badge. New lib/badge-prefs.ts;
    push-prefs Badge toggle writes localStorage, AppBadge reads it (dropped the N
    server GETs). Master notification switch was ALREADY per-device (subscribes
    this endpoint), so notifications-per-phone already works.
1b) The whole Notifications settings section is hidden outside the installed PWA
    (new useIsStandalone hook gates the AccordionItem; shared with the shell copies).
    (1c: per-KIND toggles remain account-wide — making those per-device is a
     notification-pipeline migration, flagged separately.)
2)  Investments/Investments-category divider used var(--hairline) (near-white
    #dde1e8 on the dark card) → var(--hairline-on-dark), matching every other
    settings divider.
3)  "Percent of wealth" is now an inline-edit cell like Name: renders "100%",
    tap → bare-number input, decimals with dot OR comma. ownership_share_pct
    SMALLINT → numeric(5,2) (migration 0068); zod drops .int(); the two aggregate
    share-math spots use basis points so BigInt() never sees a decimal.

TDD: aggregation-section (inline + comma decimal), app-badge + push-prefs badge
(localStorage), backend decimal-share guard (368500) — all green. tsc + lint clean.

+++ apps/api/src/routes/budget-members.ts
@@ -19,7 +19,8 @@ type MembersDeps = Pick<BootedDeps, "tenancy" | "identity">;
 
 const aggregationSchema = z.object({
   included: z.boolean(),
-  share_pct: z.number().int().min(0).max(100).optional(),
+  // Decimals allowed (e.g. 33.5%); stored as numeric(5,2).
+  share_pct: z.number().min(0).max(100).optional(),
 });
 
 export function budgetMembersRoutesFactory(deps: MembersDeps) {
